requirements.txt
flask
flask_cors
thread6
simplejson
torch
transformers
datetime
pymongo
databases
certifi
python-dotenv
TIME-python
requests
bs4
selenium
webdriver_manager
ollama
flask_jwt_extended
# gunicorn